chore: 🎸 create base class to get rid of duplicated code
